+++ common/src/main/java/dk/magnusjensen/customchestmenus/network/overlay/UpdateMenuBoundEntityS2C.java
@@ -16,31 +16,34 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package dk.magnusjensen.customchestmenus.network;
+package dk.magnusjensen.customchestmenus.network.overlay;
 
 import dk.magnusjensen.customchestmenus.Utils;
 import net.minecraft.network.RegistryFriendlyByteBuf;
-import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.ComponentSerialization;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
-import net.minecraft.resources.ResourceLocation;
 
-public record UpdateMenuTitleS2C(String menuId, Component title) implements CustomPacketPayload {
+// Handles keeping the entity ID list on the client in sync.
+public record UpdateMenuBoundEntityS2C(int entityId, OverlayOperation operation) implements CustomPacketPayload {
+    public static final CustomPacketPayload.Type<UpdateMenuBoundEntityS2C> TYPE = new CustomPacketPayload.Type<>(Utils.modLoc("add_menu_bound_entity_s2c"));
 
-    public static final CustomPacketPayload.Type<UpdateMenuTitleS2C> TYPE = new CustomPacketPayload.Type<>(Utils.modLoc("update_menu_title_s2c"));
-
-    public static final StreamCodec<RegistryFriendlyByteBuf, UpdateMenuTitleS2C> STREAM_CODEC = StreamCodec.composite(
-        ByteBufCodecs.STRING_UTF8,
-        UpdateMenuTitleS2C::menuId,
-        ComponentSerialization.STREAM_CODEC,
-        UpdateMenuTitleS2C::title,
-        UpdateMenuTitleS2C::new
+    public static final StreamCodec<RegistryFriendlyByteBuf, UpdateMenuBoundEntityS2C> STREAM_CODEC = StreamCodec.composite(
+        ByteBufCodecs.INT, UpdateMenuBoundEntityS2C::entityId,
+        StreamCodec.of(
+            (buf, operation) -> buf.writeVarInt(operation.ordinal()), // Encode enum as ordinal
+            buf -> OverlayOperation.values()[buf.readVarInt()]        // Decode ordinal back to enum
+        ), UpdateMenuBoundEntityS2C::operation,
+        UpdateMenuBoundEntityS2C::new
     );
 
     @Override
     public Type<? extends CustomPacketPayload> type() {
         return TYPE;
+    }
+
+    public enum OverlayOperation {
+        ADD,
+        REMOVE;
     }
 }