+++ common/src/main/java/dk/magnusjensen/customchestmenus/platform/services/IAttachmentHelper.java
@@ -18,9 +18,11 @@
 
 package dk.magnusjensen.customchestmenus.platform.services;
 
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.player.Player;
 
 public interface IAttachmentHelper {
-    <T> T getPlayerAttachment(ServerPlayer player, Class<T> clazz);
-    <T> void setPlayerAttachment(ServerPlayer player, T attachment);
+    <T> T getPlayerAttachment(Player player, ResourceLocation id);
+    <T> void setPlayerAttachment(ServerPlayer player, T attachment, ResourceLocation id);
 }