+++ fabric/src/main/java/dk/magnusjensen/customchestmenus/FabricClientCustomChestMenus.java
@@ -22,6 +22,7 @@ import dk.magnusjensen.customchestmenus.client.screen.CustomChestScreen;
 import dk.magnusjensen.customchestmenus.network.FabricClientNetwork;
 import dk.magnusjensen.customchestmenus.registry.FabricMenuRegistry;
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.mixin.client.rendering.WorldRendererMixin;
 import net.minecraft.client.gui.screens.MenuScreens;
 
 public class FabricClientCustomChestMenus implements ClientModInitializer {