+++ fabric/src/main/java/dk/magnusjensen/customchestmenus/FabricCustomChestMenus.java
@@ -24,10 +24,13 @@ import dk.magnusjensen.customchestmenus.registry.FabricAttachmentRegistry;
 import dk.magnusjensen.customchestmenus.registry.FabricMenuRegistry;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
+import net.fabricmc.fabric.api.entity.event.v1.ServerEntityWorldChangeEvents;
+import net.fabricmc.fabric.api.entity.event.v1.ServerPlayerEvents;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerEntityEvents;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.event.player.UseBlockCallback;
 import net.fabricmc.fabric.api.event.player.UseEntityCallback;
+import net.fabricmc.fabric.api.networking.v1.EntityTrackingEvents;
 import net.minecraft.world.InteractionResult;
 
 public class FabricCustomChestMenus implements ModInitializer {
@@ -54,6 +57,10 @@ public class FabricCustomChestMenus implements ModInitializer {
         });
 
         UseEntityCallback.EVENT.register((player, level, interactionHand, entity, entityHitResult) -> {
+            if (entityHitResult == null) {
+                return InteractionResult.PASS;
+            }
+
             var handled = EventHandler.onEntityRightClick(player, entity, interactionHand);
             if (handled) {
                 return InteractionResult.SUCCESS;
@@ -63,6 +70,12 @@ public class FabricCustomChestMenus implements ModInitializer {
         });
 
         ServerEntityEvents.ENTITY_UNLOAD.register((entity, serverLevel) -> EventHandler.onEntityUnload(entity));
+        EntityTrackingEvents.START_TRACKING.register((entity, serverPlayer) -> EventHandler.onPlayerStartTracking(serverPlayer, entity));
+        EntityTrackingEvents.STOP_TRACKING.register((entity, serverPlayer) -> EventHandler.onPlayerStopTracking(serverPlayer, entity));
+        ServerEntityWorldChangeEvents.AFTER_PLAYER_CHANGE_WORLD.register(((serverPlayer, origin, destination) -> {
+            EventHandler.onPlayerChangeDimension(serverPlayer, destination);
+        }));
+        ServerPlayerEvents.LEAVE.register(EventHandler::onPlayerLeaveServer);
 
 
         // Static load registries