+++ fabric/src/main/java/dk/magnusjensen/customchestmenus/mixin/LevelRendererMixin.java
@@ -1,0 +1,43 @@
+/*
+ *     Custom Chest Menus, a Minecraft mod that allows servers to create custom chest menus.
+ *     Copyright (c) 2025  legenden (MagnusHJensen)
+ *
+ *     This program is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This program is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package dk.magnusjensen.customchestmenus.mixin;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.mojang.blaze3d.vertex.PoseStack;
+import dk.magnusjensen.customchestmenus.client.MenuHighlightRenderer;
+import net.minecraft.client.renderer.LevelRenderer;
+import net.minecraft.client.renderer.OutlineBufferSource;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Slice;
+
+@Mixin(LevelRenderer.class)
+public abstract class LevelRendererMixin {
+    @WrapOperation(method = "method_62214",
+        slice = @Slice(from = @At(value = "INVOKE_STRING", target = "Lnet/minecraft/util/profiling/ProfilerFiller;popPush(Ljava/lang/String;)V", args = "ldc=submitEntities")),
+        at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/OutlineBufferSource;endOutlineBatch()V")
+    )
+    private void afterEntityRender(OutlineBufferSource instance, Operation<Void> original) {
+        original.call(instance);
+        MenuHighlightRenderer.renderLevelOverlay(
+            new PoseStack()
+        );
+    }
+}