+++ neoforge/src/main/java/dk/magnusjensen/customchestmenus/NeoforgeEventHandler.java
@@ -19,9 +19,11 @@
 package dk.magnusjensen.customchestmenus;
 
 import dk.magnusjensen.customchestmenus.events.EventHandler;
+import net.minecraft.server.level.ServerPlayer;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.event.entity.EntityLeaveLevelEvent;
+import net.neoforged.neoforge.event.entity.player.PlayerEvent;
 import net.neoforged.neoforge.event.entity.player.PlayerInteractEvent;
 
 @EventBusSubscriber(modid = Constants.MOD_ID)
@@ -40,7 +42,36 @@ public class NeoforgeEventHandler {
     }
 
     @SubscribeEvent
-    public static void onEntityLoad(EntityLeaveLevelEvent event) {
+    public static void onEntityUnload(EntityLeaveLevelEvent event) {
         EventHandler.onEntityUnload(event.getEntity());
+    }
+
+    @SubscribeEvent
+    public static void onPlayerStopTrackingEntity(PlayerEvent.StopTracking event) {
+        EventHandler.onPlayerStopTracking(event.getEntity(), event.getTarget());
+    }
+
+    @SubscribeEvent
+    public static void onPlayerStartTrackingEntity(PlayerEvent.StartTracking event) {
+        EventHandler.onPlayerStartTracking(event.getEntity(), event.getTarget());
+    }
+
+    @SubscribeEvent
+    public static void onPlayerChangeDimension(PlayerEvent.PlayerChangedDimensionEvent event) {
+        if (!(event.getEntity() instanceof ServerPlayer serverPlayer)) {
+            return;
+        }
+
+        var destination = serverPlayer.level().getServer().getLevel(event.getTo());
+
+        EventHandler.onPlayerChangeDimension(serverPlayer, destination);
+    }
+
+    @SubscribeEvent
+    public static void onPlayerLeaverServer(PlayerEvent.PlayerLoggedOutEvent event) {
+        if (!(event.getEntity() instanceof ServerPlayer serverPlayer)) {
+            return;
+        }
+        EventHandler.onPlayerLeaveServer(serverPlayer);
     }
 }