+++ CHANGELOG.md
@@ -5,3 +5,4 @@
     - `/ccm stop-binding` to exit binding mode.
     - `/ccm unbind` to start unbinding blocks or entities, where if a bound block/entity is right-clicked it will unbind the menu from the block/entity.
   - `/ccm list-bindings` Shows a chat list of all bound blocks and entities and a text element to teleport to it.
+  - `/ccm bind-overlay` Toggles an overlay, which draws boxes around blocks and entities that are bound to a menu if within 64 blocks of it.